add timer cycles

+++ prefixed_instructions.go
@@ -1,6 +1,6 @@
 package main
 
-var cbInstructionCycles = []uint8{
+var cbInstructionCycles = []int{
 	2, 2, 2, 2, 2, 2, 4, 2, 2, 2, 2, 2, 2, 2, 4, 2,
 	2, 2, 2, 2, 2, 2, 4, 2, 2, 2, 2, 2, 2, 2, 4, 2,
 	2, 2, 2, 2, 2, 2, 4, 2, 2, 2, 2, 2, 2, 2, 4, 2,
